fix: measure profit in terms of target coin
Prompt: Ok, the principle here is also confusing, you use USDT first but then proceed to calculate things in terms of USDC. It's very confusing. I think you should explain your logic better and then make the code clearer so that we can determine when the buy price is denominated in USDC and when it's denominated in USDT. As the target coin is not always one. We should separate the logic depending on the target coin.