fix: portfolio total calculation
That was great! But look! I dont think this is correct at all. How could we suddenly end up with a portforlio total of 21000 but the profit is only 1.3140??? Look at the total on Sonic after the first trade: 11001.32... but the price difference is supposed to be only 0.1320%. Am I missing something?